cv-extractor/Backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
pydantic
